moved Footer folder to components
