+++ frontend/src/components/Footer/_Footer.scss
